REFACTOR: split the socket server by responsibility
Replace the single 991-line implementation with three focused internal units: WebServer for exporter behavior, HTTPConnection for protocol framing, and SocketListener for transport and lifecycle. Split the 35 server tests along the same boundaries without changing their 484 assertions.
